requirements.txt
# Основные зависимости
openai==0.28.1          # Старая стабильная версия библиотеки OpenAI API
python-dotenv==1.0.1    # Загрузка переменных окружения из .env файла
flask==3.0.2           # Веб-фреймворк для создания интерфейса
Flask-SQLAlchemy==3.1.1 # ORM для работы с базой данных
Flask-Login==0.6.3     # Аутентификация пользователей
werkzeug==3.0.1        # Утилиты для Flask
Flask-WTF==1.2.1       # Работа с формами

# Версии Python
# Python >= 3.7.1 